+++ RELEASE_NOTES.md
@@ -5,6 +5,13 @@ A new header is inserted each time a *tag* is created.
 
 ## v1.12.12 (currently main branch)
 
+- Android: Gradle configuration caching is now enabled
+- Android: Filament's Gradle properties have all be renamed to `com.google.android.filament.xxx`
+  where `xxx` is the property name. See `android/build.gradle` for a complete list [⚠️]
+- Android: The Gradle property `filament_tools_dir` (now called
+  `com.google.android.filament.tools-dir`) does not have a default value anymore. Please specify
+  one in your `gradle.properties` if you reuse the Gradle plugin in your projects  [⚠️]
+
 ## v1.12.11
 
 ## v1.12.10